requirements.txt
numpy
pillow
sklearn
